requirements.txt
Flask==2.0.3
Jinja2==3.1.1
gunicorn==20.0.4
google-cloud-storage==2.3.0
sqlalchemy
google-cloud-secret-manager
google-cloud-logging
psycopg2
google-auth==1.32.1
pandas==1.3.0
pyarrow==6.0.1